+++ Source/InspectorLIFT/Public/InspectorLIFT.h
@@ -1,0 +1,25 @@
+// Copyright Epic Games, Inc. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Modules/ModuleManager.h"
+
+class FToolBarBuilder;
+class FMenuBuilder;
+
+class FInspectorLIFTModule : public IModuleInterface
+{
+public:
+    /** IModuleInterface implementation */
+    virtual void StartupModule() override;
+    virtual void ShutdownModule() override;
+
+    /** This function will be bound to Command. */
+    void PluginButtonClicked();
+
+private:
+    void RegisterMenus();
+private:
+    TSharedPtr<class FUICommandList> PluginCommands;
+};
