+++ Source/InspectorLIFTRuntime/Private/InspectorLIFTRuntime.cpp
@@ -1,0 +1,17 @@
+// Copyright Epic Games, Inc. All Rights Reserved.
+
+#include "InspectorLIFTRuntime.h"
+
+#define LOCTEXT_NAMESPACE "FInspectorLIFTRuntimeModule"
+
+void FInspectorLIFTRuntimeModule::StartupModule()
+{
+}
+
+void FInspectorLIFTRuntimeModule::ShutdownModule()
+{
+}
+
+#undef LOCTEXT_NAMESPACE
+
+IMPLEMENT_MODULE(FInspectorLIFTRuntimeModule, InspectorLIFTRuntime)
