+++ Source/InspectorLIFTRuntime/Private/Settings/IPF_SettingsContainer.cpp
@@ -1,0 +1,4 @@
+// LIFT data Inspector. All rights reserved.
+
+#include "Settings/IPF_SettingsContainer.h"
+
