--- Source/InspectorPlayFiRuntime/Private/InspectorPlayFiRuntime.cpp
@@ -1,17 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#include "InspectorPlayFiRuntime.h"
-
-#define LOCTEXT_NAMESPACE "FInspectorPlayFiRuntimeModule"
-
-void FInspectorPlayFiRuntimeModule::StartupModule()
-{
-}
-
-void FInspectorPlayFiRuntimeModule::ShutdownModule()
-{
-}
-
-#undef LOCTEXT_NAMESPACE
-
-IMPLEMENT_MODULE(FInspectorPlayFiRuntimeModule, InspectorPlayFiRuntime)
