--- Source/InspectorPlayFiRuntime/Private/Settings/IPF_SettingsContainer.cpp
@@ -1,4 +1,0 @@
-// PlayFi data Inspector. All rights reserved.
-
-#include "Settings/IPF_SettingsContainer.h"
-
