--- Source/InspectorPlayFiRuntime/Public/InspectorPlayFiRuntime.h
@@ -1,14 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "Modules/ModuleManager.h"
-
-class FInspectorPlayFiRuntimeModule : public IModuleInterface
-{
-public:
-    /** IModuleInterface implementation */
-    virtual void StartupModule() override;
-    virtual void ShutdownModule() override;
-};
